Refactor share link component
Share links are becoming a more complex concept as we prepare for
unlisted shares; this change reorganizes the code related to shares to
make it easier to modify over time.
