+++ src/app/file-browser/components/sharing-dialog/sharing-dialog.component.html
@@ -465,4 +465,9 @@
 			</div>
 		</div>
 	}
+	<pr-share-link-settings
+		[shareItem]="shareItem"
+		[shareLink]="shareLink"
+		*ngIf="canShare"
+	></pr-share-link-settings>
 </div>
