+++ src/app/file-browser/components/sharing-dialog/sharing-dialog.component.ts
@@ -1,10 +1,4 @@
-import {
-	Component,
-	ElementRef,
-	Inject,
-	OnInit,
-	ViewChild,
-} from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import {
 	UntypedFormBuilder,
 	UntypedFormGroup,
@@ -15,7 +9,6 @@ import { RelationshipService } from '@core/services/relationship/relationship.se
 import { ShareVO, ShareByUrlVO, ItemVO, ArchiveVO, InviteVO } from '@models';
 import { AccessRoleType } from '@models/access-role';
 import { sortShareVOs } from '@models/share-vo';
-import { Deferred } from '@root/vendor/deferred';
 import { DIALOG_DATA, DialogRef } from '@angular/cdk/dialog';
 import {
 	ngIfScaleAnimation,
@@ -26,20 +19,13 @@ import { AccountService } from '@shared/services/account/account.service';
 import { ApiService } from '@shared/services/api/api.service';
 import { InviteResponse } from '@shared/services/api/index.repo';
 import { ShareResponse } from '@shared/services/api/share.repo';
-import { EVENTS } from '@shared/services/google-analytics/events';
-import { GoogleAnalyticsService } from '@shared/services/google-analytics/google-analytics.service';
 import { MessageService } from '@shared/services/message/message.service';
 import {
 	PromptService,
 	RELATION_OPTIONS,
 	SHARE_ACCESS_ROLE_FIELD,
 } from '@shared/services/prompt/prompt.service';
-import { getSQLDateTime } from '@shared/utilities/dateTime';
-import { copyFromInputElement } from '@shared/utilities/forms';
-import { addDays, differenceInHours, isPast } from 'date-fns';
 import { find, partition, remove } from 'lodash';
-import { faTrash } from '@fortawesome/pro-regular-svg-icons';
-import { FeatureFlagService } from '@root/app/feature-flag/services/feature-flag.service';
 
 enum Expiration {
 	Never = 'Never',
@@ -49,19 +35,6 @@ enum Expiration {
 	Year = '1 year',
 	Other = 'Other',
 }
-
-enum ExpirationDays {
-	Day = 1,
-	Week = 7,
-	Month = 30,
-	Year = 365,
-}
-
-type ShareByUrlProps =
-	| 'defaultAccessRole'
-	| 'expiresDT'
-	| 'autoApproveToggle'
-	| 'previewToggle';
 
 const EXPIRATION_OPTIONS: FormInputSelectOption[] = Object.values(
 	Expiration,
@@ -88,18 +61,9 @@ export class SharingDialogComponent implements OnInit {
 
 	public shareLink: ShareByUrlVO = null;
 
-	public previewToggle: 0 | 1 = 1;
-	public autoApproveToggle: 0 | 1 = 1;
-	public expiration: Expiration;
-	public linkDefaultAccessRole: AccessRoleType = 'access.role.viewer';
-
 	public updatingLink = false;
 	public linkCopied = false;
 	public showLinkSettings = false;
-
-	public linkType = 'private';
-
-	public trashIcon = faTrash;
 
 	public newAccessRole: AccessRoleType = 'access.role.viewer';
 	public accessRoleOptions: FormInputSelectOption[] =
@@ -111,26 +75,10 @@ export class SharingDialogComponent implements OnInit {
 	public showInvitationForm = false;
 	public invitationForm: UntypedFormGroup;
 
-	public shareLinkTypes = [
-		{
-			text: 'Anyone can view',
-			value: 'public',
-			description: 'Anyone with the link can view and download.',
-		},
-		{
-			text: 'Restricted',
-			value: 'private',
-			description: 'Must be logged in to view.',
-		},
-	];
-
-	public displayDropdown = false;
-
-	@ViewChild('shareUrlInput', { static: false }) shareUrlInput: ElementRef;
-
 	public archiveFilterFn = (a: ArchiveVO) =>
 		!find(this.shares, { archiveId: a.archiveId }) &&
 		!find(this.shares, { archiveId: a.archiveId });
+
 	constructor(
 		@Inject(DIALOG_DATA) public data: any,
 		private accountService: AccountService,
@@ -140,9 +88,7 @@ export class SharingDialogComponent implements OnInit {
 		private api: ApiService,
 		private messageService: MessageService,
 		private relationshipService: RelationshipService,
-		private ga: GoogleAnalyticsService,
 		private route: ActivatedRoute,
-		private feature: FeatureFlagService,
 	) {
 		this.invitationForm = this.fb.group({
 			fullName: ['', [Validators.required]],
@@ -152,8 +98,6 @@ export class SharingDialogComponent implements OnInit {
 		});
 
 		this.shareItem = this.data.item as ItemVO;
-
-		this.displayDropdown = feature.isEnabled('unlisted-share');
 	}
 
 	ngOnInit(): void {
@@ -173,7 +117,6 @@ export class SharingDialogComponent implements OnInit {
 		this.relationshipService.update();
 
 		this.shareLink = this.data.link;
-		this.setShareLinkFormValue();
 
 		this.checkQueryParams();
 	}
@@ -412,150 +355,6 @@ export class SharingDialogComponent implements OnInit {
 					translate: true,
 				});
 			}
-		}
-	}
-
-	getExpirationFromExpiresDT(expiresDT: string): Expiration {
-		if (!expiresDT) {
-			return Expiration.Never;
-		}
-
-		const diff = differenceInHours(
-			new Date(expiresDT),
-			new Date(this.shareLink.createdDT),
-		);
-
-		if (diff <= 24 * ExpirationDays.Day) {
-			return Expiration.Day;
-		} else if (diff <= 24 * ExpirationDays.Week) {
-			return Expiration.Week;
-		} else if (diff <= 24 * 7 * ExpirationDays.Month) {
-			return Expiration.Month;
-		} else if (diff <= 24 * 7 * ExpirationDays.Year) {
-			return Expiration.Year;
-		} else {
-			return Expiration.Other;
-		}
-	}
-
-	getExpiresDTFromExpiration(expiration: Expiration): string {
-		switch (expiration) {
-			case Expiration.Never:
-				return null;
-			case Expiration.Day:
-				return getSQLDateTime(addDays(new Date(this.shareLink.createdDT), 1));
-			case Expiration.Week:
-				return getSQLDateTime(addDays(new Date(this.shareLink.createdDT), 7));
-			case Expiration.Month:
-				return getSQLDateTime(addDays(new Date(this.shareLink.createdDT), 30));
-			case Expiration.Year:
-				return getSQLDateTime(addDays(new Date(this.shareLink.createdDT), 365));
-		}
-	}
-
-	setShareLinkFormValue(): void {
-		if (this.shareLink) {
-			this.previewToggle = this.shareLink.previewToggle;
-			this.autoApproveToggle = this.shareLink.autoApproveToggle || 0;
-			this.expiration = this.getExpirationFromExpiresDT(
-				this.shareLink.expiresDT,
-			);
-			this.linkDefaultAccessRole = this.shareLink.defaultAccessRole;
-			this.expirationOptions = EXPIRATION_OPTIONS.filter((expiration) => {
-				switch (expiration.value) {
-					case Expiration.Never:
-					case Expiration.Other:
-						return true;
-					default:
-						return !isPast(
-							new Date(
-								this.getExpiresDTFromExpiration(expiration.value as Expiration),
-							),
-						);
-				}
-			});
-		} else {
-			this.previewToggle = 1;
-			this.autoApproveToggle = 1;
-			this.expiration = Expiration.Never;
-			this.expirationOptions = EXPIRATION_OPTIONS;
-		}
-	}
-
-	async generateShareLink() {
-		this.updatingLink = true;
-		try {
-			const response = await this.api.share.generateShareLink(this.shareItem);
-			this.shareLink = response.getShareByUrlVO();
-			this.shareLink.autoApproveToggle = this.autoApproveToggle || 0;
-			this.shareLink.previewToggle = this.previewToggle || 0;
-			await this.api.share.updateShareLink(this.shareLink);
-			this.setShareLinkFormValue();
-			this.showLinkSettings = true;
-			this.ga.sendEvent(EVENTS.SHARE.ShareByUrl.initiated.params);
-		} catch (err) {
-			if (err instanceof ShareResponse) {
-				this.messageService.showError({
-					message: err.getMessage(),
-					translate: true,
-				});
-			}
-		} finally {
-			this.updatingLink = false;
-		}
-	}
-
-	copyShareLink() {
-		const element = this.shareUrlInput.nativeElement as HTMLInputElement;
-
-		copyFromInputElement(element);
-
-		this.linkCopied = true;
-		setTimeout(() => {
-			this.linkCopied = false;
-		}, 5000);
-	}
-
-	async removeShareLink() {
-		const deferred = new Deferred();
-		try {
-			await this.promptService.confirm(
-				'Remove link',
-				'Are you sure you want to remove this link?',
-				deferred.promise,
-				'btn-danger',
-			);
-
-			await this.api.share.removeShareLink(this.shareLink);
-			this.shareLink = null;
-			this.setShareLinkFormValue();
-			deferred.resolve();
-			this.showLinkSettings = false;
-		} catch (response) {
-			deferred.resolve();
-			if (response instanceof ShareResponse) {
-				this.messageService.showError({ message: response.getMessage() });
-			}
-		}
-	}
-
-	async onShareLinkPropChange(propName: ShareByUrlProps, value: any) {
-		this.updatingLink = true;
-		try {
-			const update = new ShareByUrlVO(this.shareLink);
-			update[propName] = value;
-			await this.api.share.updateShareLink(update);
-			this.shareLink[propName] = update[propName];
-		} catch (err) {
-			if (err instanceof ShareResponse) {
-				this.messageService.showError({
-					message: err.getMessage(),
-					translate: true,
-				});
-			}
-			this.setShareLinkFormValue();
-		} finally {
-			this.updatingLink = false;
 		}
 	}
 }
